# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: PrintFlow 3D API
  description: |
    API REST de PrintFlow — Gestion d'impressions 3D et filaments.

    ## Authentification
    Si l'authentification est activée dans les paramètres, ajoutez le header :
    ```
    X-Auth-Token: votre_mot_de_passe
    ```

    ## Intégration Home Assistant
    Exemple de sensor REST pour le stock de filaments :
    ```yaml
    sensor:
      - platform: rest
        name: PrintFlow stock faible
        resource: http://192.168.1.145/api/stats/alerts
        method: GET
        value_template: "{{ value_json | length }}"
        unit_of_measurement: "bobines"
    ```

  version: "1.7.0"
  contact:
    name: PrintFlow 3D
    url: https://github.com/Fabdub59650/printflow-3d
servers:
  - url: /api
    description: Serveur local (Raspberry Pi)
...
paths:
...
  /filaments:
    get:
      tags: [Filaments]
      summary: Lister les filaments
      parameters:
        - name: archived
          in: query
          description: "1 pour inclure les bobines archivées"
          schema:
            type: integer
            enum: [0, 1]
      responses:
        "200":
          description: Liste des filaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filament'
              example:
                - id: 1
                  name: "PLA Blanc Elegoo"
                  brand: "Elegoo"
                  material: "PLA"
                  color_hex: "#FFFFFF"
                  color_name: "Blanc"
                  weight_remaining: 850
                  weight_total: 1000
                  temp_nozzle_min: 190
                  temp_nozzle_max: 220
                  archived: 0
    post:
      tags: [Filaments]
      summary: Créer un filament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilamentInput'
            example:
              name: "PETG Noir"
              brand: "Bambu"
              material: "PETG"
              color_hex: "#000000"
              color_name: "Noir"
              weight_total: 1000
              weight_remaining: 1000
              temp_nozzle_min: 230
              temp_nozzle_max: 250
      responses:
        "201":
          description: Filament créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filament'
  /filaments/{id}:
    get:
      tags: [Filaments]
      summary: Détail d'un filament
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        "200":
          description: Détail du filament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filament'
    put:
      tags: [Filaments]
      summary: Modifier un filament
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilamentInput'
      responses:
        "200":
          description: Filament mis à jour
    delete:
      tags: [Filaments]
      summary: Supprimer un filament
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        "200":
          description: Filament supprimé
  /filaments/import-csv:
    post:
      tags: [Filaments]
      summary: Importer des filaments depuis un CSV
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              description: Contenu CSV avec en-têtes
      responses:
        "200":
          description: Import réussi
          content:
            application/json:
              example:
                imported: 5
                skipped: 1

  # ── IMPRESSIONS ───────────────────────────────────────────
...
  /maintenance:
    get:
      tags: [Maintenance]
      summary: Lister les entrées de maintenance
      responses:
        "200":
          description: Liste des maintenances
          content:
            application/json:
              example:
                - id: 1
                  printer_name: "Elegoo Centauri"
                  type: "nettoyage"
                  performed_at: "2026-04-01"
                  next_due: "2026-05-01"
                  description: "Nettoyage buse et plateau"
    post:
      tags: [Maintenance]
      summary: Créer une entrée de maintenance
      requestBody:
        required: true
        content:
          application/json:
            example:
              printer_id: 1
              type: "nettoyage"
              performed_at: "2026-04-08"
              next_due: "2026-05-08"
              description: "Nettoyage complet"
      responses:
        "201":
          description: Maintenance créée
  /maintenance/{id}:
    put:
      tags: [Maintenance]
      summary: Modifier une entrée de maintenance
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            example:
              next_due: "2026-06-01"
      responses:
        "200":
          description: Maintenance mise à jour
    delete:
      tags: [Maintenance]
      summary: Supprimer une entrée de maintenance
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        "200":
          description: Maintenance supprimée
  /maintenance/{id}/done:
    patch:
      tags: [Maintenance]
      summary: Marquer une maintenance comme effectuée
      description: Efface la date de prochaine échéance (retire l'alerte du dashboard)
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        "200":
          description: Maintenance marquée comme effectuée

  # ── STATISTIQUES ──────────────────────────────────────────
...
  /history:
    get:
      tags: [Historique]
      summary: Journal des modifications
      parameters:
        - name: entity
          in: query
          description: Filtrer par type d'entité
          schema:
            type: string
            enum: [filament, print, printer, project, library_object, nfc]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Entrées du journal
          content:
            application/json:
              example:
                - id: 1
                  entity: "filament"
                  entity_id: 3
                  action: "create"
                  detail: "Filament créé : PLA Rouge"
                  created_at: "2026-04-08T10:30:00Z"

  # ── NFC ───────────────────────────────────────────────────
  /nfc/history/{filamentId}:
    get:
      tags: [NFC]
      summary: Historique des écritures NFC d'un filament
      parameters:
        - name: filamentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Historique des écritures NFC
          content:
            application/json:
              example:
                - id: 1
                  uid: "04A3B2C1D0E5F6"
                  format: "elegoo"
                  material: "PLA"
                  subtype: "PLA+"
                  color_hex: "#FFFFFF"
                  written_at: "2026-04-08T09:15:00Z"

  # ── PARAMÈTRES ────────────────────────────────────────────
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    Filament:
      type: object
      properties:
        id:          { type: integer }
        name:        { type: string }
        brand:       { type: string }
        material:    { type: string, enum: [PLA, PETG, ABS, ASA, TPU, Nylon, PC, HIPS, PVA, autre] }
        color_hex:   { type: string, example: "#3b82f6" }
        color_name:  { type: string }
        diameter:    { type: number, example: 1.75 }
        weight_total:     { type: number }
        weight_remaining: { type: number }
        spool_weight:     { type: number }
        temp_nozzle_min:  { type: integer }
        temp_nozzle_max:  { type: integer }
        temp_bed_min:     { type: integer }
        temp_bed_max:     { type: integer }
        price:       { type: number }
        location:    { type: string }
        notes:       { type: string }
        nfc_uid:     { type: string }
        archived:    { type: integer, enum: [0, 1] }
    FilamentInput:
      type: object
      required: [name, material]
      properties:
        name:        { type: string }
        brand:       { type: string }
        material:    { type: string }
        color_hex:   { type: string }
        color_name:  { type: string }
        diameter:    { type: number }
        weight_total:     { type: number }
        weight_remaining: { type: number }
        temp_nozzle_min:  { type: integer }
        temp_nozzle_max:  { type: integer }
        temp_bed_min:     { type: integer }
        temp_bed_max:     { type: integer }
        price:       { type: number }
        location:    { type: string }
        notes:       { type: string }
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: Requis uniquement si l'authentification est activée dans les paramètres
security:
  - ApiKeyAuth: []
  - {}